Added diff check, it needs to implement a better visulisation of the difference and remove the prints.
List needs to have the format changed.

Regaining connection should not message unless the connection was lost
for long enough for a notification to have been sent/printed.

Functions could be added to a lot of the duplicate content I just have
to decide how I want to implement them.

A lot of the names could be improved to better representations of their
function.